ci(ansible): add new workflow to support running ansible on a schedule
There was partial support for this, but it wasn't working because the workflow contains logic to skip running. but for the scheduled run it should always execute. In order to over come this, the scheduled one is now it's own workflow and is a re-usable workflow and executes on a schedule. The original one that runs as part of pipeline runs for branches, etc still has the logic and now executes the new workflow(based on the branch)